+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.7.44
+
+* add `get_timestamp` to `ExpiringEntryDict`